update demo parser

+++ dags/lambda_codes/ingest/parser/demo_parser.py
@@ -73,5 +73,8 @@ def delete_local_tmp(tempdirectory):
     files_to_delete = os.listdir(tempdirectory)
     print('deleting files: ',files_to_delete)
     for single_file in files_to_delete:
-        filepath = os.path.join(tempdirectory, single_file)
-        os.remove(filepath)
+        filepath = os.path.join(directory, single_file)
+        if 'go' in filepath:
+            pass
+        else:
+            os.remove(filepath)
